+++ spec/requests/as_account/playlist_spec.rb
@@ -180,8 +180,10 @@ describe Yt::Playlist, :device_app do
     let(:id) { $account.channel.playlists.first.id }
 
     it 'returns valid reports for playlist-related metrics' do
+      expect{playlist.views}.not_to raise_error
       expect{playlist.playlist_starts}.not_to raise_error
 
+      expect{playlist.views_on 3.days.ago}.not_to raise_error
       expect{playlist.playlist_starts_on 3.days.ago}.not_to raise_error
     end
   end