+++ spec/requests/as_content_owner/playlist_spec.rb
@@ -10,6 +10,53 @@ describe Yt::Playlist, :partner do
     context 'managed by the authenticated Content Owner' do
       let(:id) { ENV['YT_TEST_PARTNER_PLAYLIST_ID'] }
 
+      describe 'views can be retrieved for a specific day' do
+        context 'in which the playlist was viewed' do
+          let(:views) { playlist.views_on ENV['YT_TEST_PARTNER_PLAYLIST_DATE']}
+          it { expect(views).to be_a Float }
+        end
+
+        context 'in which the playlist was not viewed' do
+          let(:views) { playlist.views_on 20.years.ago}
+          it { expect(views).to be_nil }
+        end
+      end
+
+      describe 'views can be retrieved for a range of days' do
+        let(:date) { 4.days.ago }
+
+        specify 'with a given start (:since option)' do
+          expect(playlist.views(since: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:until option)' do
+          expect(playlist.views(until: date).keys.max).to eq date.to_date
+        end
+
+        specify 'with a given start (:from option)' do
+          expect(playlist.views(from: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:to option)' do
+          expect(playlist.views(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'views can be grouped by day' do
+        let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
+        let(:keys) { range.values }
+
+        specify 'without a :by option (default)' do
+          views = playlist.views range
+          expect(views.keys).to eq range.values
+        end
+
+        specify 'with the :by option set to :day' do
+          views = playlist.views range.merge by: :day
+          expect(views.keys).to eq range.values
+        end
+      end
+
       describe 'playlist starts can be retrieved for a specific day' do
         context 'in which the playlist was viewed' do
           let(:playlist_starts) { playlist.playlist_starts_on ENV['YT_TEST_PARTNER_PLAYLIST_DATE']}
@@ -62,6 +109,7 @@ describe Yt::Playlist, :partner do
       let(:id) { 'PLbpi6ZahtOH6J5oPGySZcmbRfT7Hyq1sZ' }
 
       specify 'playlist starts cannot be retrieved' do
+        expect{playlist.views}.to raise_error Yt::Errors::Forbidden
         expect{playlist.playlist_starts}.to raise_error Yt::Errors::Forbidden
       end
     end